requirements.txt
anyascii==0.3.2
asgiref==3.6.0
beautifulsoup4==4.11.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
distlib==0.3.7
Django==4.2
django-admin-chat==0.6
django-admin-volt==1.0.10
django-appconf==1.0.5
django-cors-headers==4.2.0
django-crontab==0.7.1
django-filter==23.2
django-flexible-subscriptions==0.15.1
django-fsm==2.8.1
django-fsm-log==3.1.0
django-jsonfield==1.4.1
django-modeladmin-reorder==0.3.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==4.0.0
django-tinymce==3.6.1
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
drf-extra-fields==3.4.1
et-xmlfile==1.1.0
filelock==3.12.4
filetype==1.2.0
firebase-admin==6.2.0
geographiclib==2.0
geopy==2.4.0
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
html5lib==1.1
httplib2==0.22.0
idna==3.4
l18n==2021.3
Markdown==3.4.3
msgpack==1.0.5
numpy==1.25.1
openpyxl==3.1.2
Pillow==9.5.0
pillow-heif==0.13.0
platformdirs==3.10.0
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyotp==2.8.0
pyparsing==3.1.0
pytz==2023.3
razorpay==1.4.1
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.5
webencodings==0.5.1
Willow==1.6.1
django-xhtml2pdf==0.0.4